Full system test as a DAG pipeline
This demonstrates how the full system test can be built
and run as a DAG pipeline.

We separate description of the workflow and runtime.

One of the benefits coming with this are
automatic parallel scheduling of QED + normal simulation and
parallel RAW data creation (as long as this is possible in memory).

Makes use of a recent development prototyped in O2DPG
https://github.com/AliceO2Group/O2DPG/blob/master/MC/doc/WorkflowRunner.md

The commit should be seen as a first demonstrator rather than a full conversion.