[MCH] avoid using HB packets to compute digits time
The computation of the Digits time is now based on the
orbits information in each Time Frame.

The SAMPA BC counter at the beginning of a given TF is
estimated by counting the total number of orbits elapsed
since the beginning of the run.

The computation assumes that the SAMPA BC counters are
simultaneously reset via a dedicated trigger signal
some time before the beginning of the run.