MCHRaw: Introduce a few new types to clarify some interfaces.
In particular the channel was ambiguous as it could be a sampa channel
(0..31) or a dual sampa channel (0..63).

Also moved a few template definitions from .h to .cxx to fix clang-tidy
misc-definitions-in-headers warning (to avoid potential ODR violations).

And also add a dummy-elecmap option to the DataDecoder (to be able to test
decoding on the full detector, not only the part that is currently
installed and operational)