[EMCAL-728] Reduce verbosity of the error logging
Errors related to raw channels (HW channels, DDL,
cell ID are cached in a std::unordered_map. In case
the same message has already been printed before
the message is suppressed in all further occurrences
and instead the number of occurrences is counted.

In addition option to switch off merging of high gain/
low gain cells, this is particular of relevance for MC
where only either of the two channel types are stored,
resulting in a large amount of raw data errors.