Optionally pass to any EntropyEncoders memory alloc.margin factor
If the detector data is very different from the one used for the entropy encoding dictionary creation, the
compressed data size may exceed the uncompressed one. In this case the standard estimate of the buffer to
allocate for the compressed data might be insufficient and an exception will be produced.
o2-<det>-entropy-encoder-workflow --mem-factor <F>
will apply a factor F (e.g. 1.5) to the standard allocation policy.