DPL: Adjusting the InputSpan backend
InputRecord is providing all payloads as header payload pairs, even
if there are multiple payloads per header. The InputSpan backend is
using a specific API of messageSet to access all payloads in
header-payload pairs

This is following the original O2 data model. As an extension, the
API allows to access also the payloads of a sequence with multiple
payloads per header in header-payload pairs.