Extending DPL input handling and message caching related to [O2 1325]  (#7110)
This is supporting an upcoming extension of the O2 data model to allow
not only header-payload pairs but also sequences of payloads preceded
by a single header message. Such a sequence can be indicated by setting
splitPayloadIndex == splitpayloadParts with splitpayloadParts > 1.
A change in the version of DataHeader will be discussed, although there
is no change in the members of the header.

There is no change on the InputRecord API, only the payload size of the
current payload is provided in the DataRef object. It needs to be discussed,
which payload size is going to be provided in the data header in case of
a sequence with multiple payloads.

Changes alongside (most of them meanwhile moved to separate commits):
- make MessageSet the owner of FairMQMessage objects in a linear vector,
  enhances performance also for split payload sequences in header-paiload
  format
- generalizing DataRelayer to relay array of messages
- storing the payload message size in the DataRef

Extending and cleaning up DataRelayer test and benchmark
- allocating header stack through transport allocator and avoiding copy
- adding test for split parts in the header-payload scheme
- adding test for split parts in payloads sequence scheme

During the development a performance regression has been observed when
using o2::pmr::getMessage, needs to be investigated.