Pass-trough mode for PVertexer, use it in cosmic
Since the aod creation is based on the vertex-tracks associations, the vertexer must
always run, even if there is no chance to find any vertex: in this case a special
VtxTrackRef record for non-associated tracks will be created.
Therefore the pvertexer is enabled in the dpl-workflow.sh even in absence of the ITS.
In the cosmic mode a special pass-through mode is deployed, which will simply create
empty vertex output.