[EMCAL-630] Correction for BC mod
Due to clock difference between LHC (40 MHz) and
ALTRO (10 MHz) 4 bunch crossings are within one
ALTRO sample. Consequently the fitted time of the trigger
of the collision for the trigger BC is shifted by 25, 50 and 75
ns for the different BC % 4 (and consequently also
translates into the pileup peaks). In order to align the time
peaks for the differnt trigger BC the convention is used to
shift the times for trigger in BC % 4 != 0 to BC %4 = 0.