Option to fowards MC tracks from O2-sim to external (DPL) consumer
The hit merger can now optionally forward kinematics (MCEventHeaders, MCTracks)
on a FairMQ channel. This is useful for on-the-fly processing of tracks
in analysis or other (DPL) consumers.

On the DPL side, a new proxy device is offered to pickup and forward
MCTracks and MCEventHeaders on DPL channels.

This PR offers prototype status to play with the new feature. See
the new SimExample for instructions.

A couple of things remain to be done:
a) configuration of the FairMQ channel on which mctracks are sent
   (currently works for one instance of o2-sim only)
b) ability to use pipes to connect o2-sim, the proxy, and further consumers:
   `o2-sim ... | o2-sim-mctracks-proxy | some-analysis`

https://alice.its.cern.ch/jira/browse/O2-3319