Dummy driver device with no input and dummy output
Sometimes we need a workflow to execute some number of loops with the timing info
injected according to some policy, e.g. to fetch certain ccdb objects via DPL fetcher.
In this case this device can be added as an upstream one in the workflow equiped with the
HBFUtilsInitializer stuff which allow to steer timing info via dynamically added
--hbfutils-config option

For the real example of usage see barrel-aligment-workflow.cxx: it allows to
use it as a simple executable w/o real input TF data but with the proper timing.