Fix in the VDrift selection logic
In case the current TF timestamp is < then that of the previous TF (e.g. due to loading another CTF file),
the vdrift of previous one was used even if another object was supplied by the CCDB fetcher
(since the creation time of previous object is > than that of the newly supplied one).

This PR ensures the proper selection of updated object among different VDrift sources.