[O2-3142] DPL: fire uv_timer in smaller intervals
This is an attempt to fix the issue of poor randomness of the cycle shifts of QC tasks on EPNs.

It appears that the uv_timer itself is actually never shifted, only the *last variable in
`timeDrivenCreation` callback is. Thus, the two timers are out of sync and in the end we
generate the timer input only once the uv_timer fires. The problem did not appear with DPL driver,
because the main loop was being triggered due to different reasons (e.g. WebSocket events), which
do not take place when a data processor is executed on its own, like in ODC-controlled environment.

This is a rather crude way to fix it by increasing the period of uv_timer by `resolution` and relying
on the `timeDrivenCreation` callback to check if enough time elapsed. I confirmed that the CPU usage with
a cycle of 10s stays below 1%, so the performance impact seems to be negligible. However, a nicer
solution would be probably to control the timer from within the `timeDrivenCreation` callback and
somehow detect when this timer has fired (but not the others).