Optionally ask ccdb-populator to validate upload
If the ccdb-populator is called with the command line option validate-upload,
or the CcdbObjectInfo for the object was prepared with setValidateUpload(true),then
for every uploaded object it will read back its headers and will make sure that
they correspond to an object with requested validity ranges and uploaded < 3 secods ago.

In case of failure an ERROR (or a FATAL if --fatal-on-failure was passed on the command line)
will be logged, on success: an IMPORTANT level message (shown as INFO in the logger)
Validated upload ... will be logged.