DPL: fix the cpu_usage_fraction metric
We need to add up time, we cannot consider the absolute value like
we do for the transmitted bytes.

Notice also that the value is in general between 0 and 1000 but it
can be > 1000 due to rounding errors (i.e. some processing time from
the previous interval assigned to the current one).