o2-sim: Stop tracks when they reach max time of flight
We stop track propagation when the time of flight limit is reached.
(Somehow this does not appear to be working correctly by just
 using the kTOFMAX cut value).

Set the default TOFMAX to 0.1s (instead of 1E10 seconds) which should be enough
to capture afterglow effects in TPC etc. Still it prevents
hits being created sometimes hours after the event, which in turn leads
to crashes in some digitizers.

The commit is only relevant for certain studies, for example
when low energy neutron physics is switched on.