[EMCAL-1046] Swap HG and LG at overflow cut
The reconstruction swap HG and LG at 950 ADC counts
in order to avoid saturation of the HG (i.e. due to
pedestals). The compression instead used the 10 bit
ADC range (1024). In the encoding LG cells in that
energy range are encoded with negative values and
therefore are out-of-range. In addition introduce
a 200 MeV safety margin for the LG in order to not
truncate for cells which are slightly below the
saturation offset due to floating point imprecision.