[EMCAL-1048] Treat faulty memory size reported by RDH
In case the memory size in the RDH reported is larger
then the raw payload size forward the current positon
to payload size or offsetToNext - whatever is smaller,
before raising the error.

In case the trailer is incomplete the page cannot be trusted,
likely because it is overwritten in memory. The marker must
be forwarded and an exception is thrown.