TRD raw data display  (#11552)
* Import classes to analyse TRD digits, tracklets

The classes allow to load trdtracklets.root and trddigits.root files,
iterate over digits and tracklets and partition the data by MCM.

* Add coordinate transformer and DrawMCM routine

* Consider ExB in coordinate transformer

* Add display of cluster positions

* Add method to iterate over RawDataSpan by padrow

* Make raw display compatible with real data

Fix issue when no MC hits file is found.
Switch off ploting of clusters by default to make plots cleaner.

* Split RawDisplay code into 3 files

The splitting prepares the code for inclusion into O2. The functionality is
essentially the same as before.

* Move TRD raw data display to QC subdirectory

This move prepares the raw data display and analysis utitlities for
inclusion in the upstream repository.

The classes also move out of the o2::trd::rawdisp into the o2::trd
namespace. The DataManager class was renamed RawDataManager.

* Clean up and add comments

* Add drawing options for RawDisplay

* Add example macro

* Run clang-format and add copyright message

* Run clang-format again

* Add test for drawing demo macro

* Work on Monte-Carlo hits

* Move helper structs from RawDataManager.h to .cxx

* Implement visualisation of hits

* Apply clang-format

* Fix formatting

* Clean up output / logging

* Use PadPlane for coordinate transformation

* Fix formating

* Convert RawDataSpan to use boot:iterator_range

* Implement drawing of MC track segments

* Fix raw display example macro

* Fix formatting

* remove TTreeReaderArray and fix a bug

---------

Co-authored-by: Sean Murray <hamiltonthomas@gmail.com>