FT0: Use dead channel map in digitizer (#12340)
* FT0: Use dead channel map in digitizer

Enables the use of the dead channel map from CCDB in the FT0 digitizer as default.
Hits in dead channels will not be processed. To disable checking the dead channel map,
pass `--FT0Digitizer --disable-dead-channel-map` to the digitizer workflow.

* FDD: Use dead channel map in digitizer

Enables the use of the dead channel map from CCDB in the FDD digitizer as default. Hits in dead channels will not be processed. To disable checking the dead channel map, pass --FDDDigitizer --disable-dead-channel-map to the digitizer workflow.