o2-sim: Change way to create standalone MCHeader file
This is changing the way we create o2sim_MCHeader.root, which
is now created directly in the merger device.

The reason for this change is an (unexplainable) segmentation fault in the routine
MCEventHeader::extractFileFromKinematics on some ARM servers (CNAF).
This routine was until now used to copy MC header information from
o2sim_Kine.root to o2sim_MCHeader.root

The present change avoids this segfault.