Utility to keep only some trees out of the AO2Ds (#13422)
Conveniently called strainer, to avoid confusion with skimming, thinning, dicing, molding, and all the other manufacturing terms used for our data processing
Implement a random downsampling feature to reduce the size of the output
AO2D file. This allows users to control the output file size, which can be
useful when working with large datasets. The downsampling factor is
specified as a command-line argument and is used to randomly select a
subset of the input data to include in the output file.