AggregatedRunInfo struct proposal
We are often in the need to collect global properties of a run like
run-duration, run-start, firstOrbit, lastOrbit, ...

No single CCDB source delivers all of these information. Moreover we sometimes
have multiple sources for the same information (GRPECS, RCT/Info/RunInformation)
which may not be at sync.

The proposal of this commit is to:

(a) introduce an aggregator struct "AggregatedRunInfo" in which we can
    deliver all important information describing a Run.

(b) implement an authoritative and agreed-on method to fill this struct.

The goal is to set everyone (reco, MC, analysis) on the same foot/contract.

For now the struct is minimal for demonstration purposes.