Fix fake/correct status for barrel tracks
The fMcMask bit 15 (fake global track label or TOF_label != TPC_lable) was
wrong since original TOF cluster label (set in the reconstruction) was compared
with TPC remapped label prepared for AOD storage.
In fact, we don't need to consider separately the global_label.isFake and TOF-TPC mismach:
TOF is the last detector in the matching process and the global track label is determined
by the TPC track label. Hence, if the TOF match is present but its cluster is not contributed
by the TPC track, the global label isFake will be necessarilly true, and vice versa.

Also, the status of bit 13 (flagging ITS-TPC mismatch) was covering only track-to-track
matches but not those from the afterburner.

Now settings of fakeness relies on the isFake status from the reconstruction.