[MUON] use track time in global muon matching
Added option for using the track time when selecting the MFT candidates
to be matched with a given MCH track.
This helps to reduce the number of candidates when the MCH track is
already matched with MID and therefore has a precise timing.
In this case the number of MCH tracks that need to be combined with
MFT tracks from two adjacent ROFs is strongly reduced.