DPL Analysis: workaround to publish histograms also with pipelining (#13980)
For some reason if the histograms arrive all at once, they get dropped.
Not yet sure why that happens. It clearly cannot merely be a matter of
"older possible timeframe" being wrong, nor a problem with the order of
the end of stream, because otherwise I would expect also this to fail.