[MCH] enable mapping fixes in ASYNC mode
The MCHDigitModifier.updateST1=true and
MCHDigitModifier.updateST2=true options are added to
the MCH reco workflow for the ASYNC processing of
real data.
This enables the code that fixes the readout mapping
in existing CTFs for some parts of ST1 and ST2 detectors.

The mapping has also been corrected in the O2 code.
Therefore the remapping is disabled for SYNC processing
and MC simulations, since they already apply the correct
mapping to generate the digits.