Improve exit handling of O2HitMerger
So far, the exit status of O2HitMerger was not analysed.
This could lead to situations where O2HitMerger was killed by the OS
due to out-of-memory, yet the o2-sim simulator still exited as "successfull".

This commit improves the handling. Problems in O2HitMerger will lead
to exit code 1 of o2-sim.