Add to trackdata in residuals TOF time difference and t0 error
TrackData got extra data member deltaTOF, which provides the difference between
the measered and expected time-of-flight in ps: timeTOF - t0 - tof_integral(track_PID)
if the the tracks was matched to TOF (0 otherwhise).
The meaning of the clAvailTOF data member is changed: as before, 0 means no TOF match,
otherwhise the value of the uncertainty on the interaction t0 is stored in ps (200 means
that the nominal BC compatible with the TOF time was assumed).